+++ lib/scout_apm/environment.rb
@@ -50,6 +50,10 @@ module ScoutApm
         '.'
       end
     end
+
+    def heroku?
+      ENV['dyno']      
+    end
     
     # This needs to be improved. Frequently, multiple app servers gem are present and which
     # ever is checked first becomes the designated app server.